+++ pkg/lightning/mydump/loader.go
@@ -37,13 +37,14 @@ type MDDatabaseMeta struct {
 }
 
 type MDTableMeta struct {
-	DB         string
-	Name       string
-	SchemaFile FileInfo
-	DataFiles  []FileInfo
-	charSet    string
-	TotalSize  int64
-	IndexRatio float64
+	DB           string
+	Name         string
+	SchemaFile   FileInfo
+	DataFiles    []FileInfo
+	charSet      string
+	TotalSize    int64
+	IndexRatio   float64
+	IsRowOrdered bool
 }
 
 type SourceFileMeta struct {
@@ -428,12 +429,13 @@ func (s *mdLoaderSetup) insertTable(fileInfo FileInfo) (*MDTableMeta, bool, bool
 	}
 	s.tableIndexMap[fileInfo.TableName] = len(dbMeta.Tables)
 	ptr := &MDTableMeta{
-		DB:         fileInfo.TableName.Schema,
-		Name:       fileInfo.TableName.Name,
-		SchemaFile: fileInfo,
-		DataFiles:  make([]FileInfo, 0, 16),
-		charSet:    s.loader.charSet,
-		IndexRatio: 0.0,
+		DB:           fileInfo.TableName.Schema,
+		Name:         fileInfo.TableName.Name,
+		SchemaFile:   fileInfo,
+		DataFiles:    make([]FileInfo, 0, 16),
+		charSet:      s.loader.charSet,
+		IndexRatio:   0.0,
+		IsRowOrdered: true,
 	}
 	dbMeta.Tables = append(dbMeta.Tables, ptr)
 	return ptr, dbExists, false
@@ -444,11 +446,12 @@ func (s *mdLoaderSetup) insertView(fileInfo FileInfo) (bool, bool) {
 	_, ok := s.tableIndexMap[fileInfo.TableName]
 	if ok {
 		meta := &MDTableMeta{
-			DB:         fileInfo.TableName.Schema,
-			Name:       fileInfo.TableName.Name,
-			SchemaFile: fileInfo,
-			charSet:    s.loader.charSet,
-			IndexRatio: 0.0,
+			DB:           fileInfo.TableName.Schema,
+			Name:         fileInfo.TableName.Name,
+			SchemaFile:   fileInfo,
+			charSet:      s.loader.charSet,
+			IndexRatio:   0.0,
+			IsRowOrdered: true,
 		}
 		dbMeta.Views = append(dbMeta.Views, meta)
 	}
